Refactoring: A little optimize the code, create dependencies file
