+++ src/backend/database.py
@@ -15,7 +15,3 @@ AsyncSessionLocal = sessionmaker(bind=engine, class_=AsyncSession, expire_on_com
 
 Base = declarative_base()
 from . import models
-
-async def get_db():
-    async with AsyncSessionLocal() as session:
-        yield session
